--- Contributing.md
@@ -1,42 +1,0 @@
-# Руководство по внесению изменений
-
-Поддерживайте ваш репозиторий обновлённым. Когда наставник принимает ваш пулреквест, он попадает в репозиторий Академии, но не в ваш форк.
-
-#### 1. Не коммитьте ничего самостоятельно в `master` вашего репозитория
-
-Это помешает вам аккуратно обновлять ваш репозиторий, могут возникнуть конфликты.
-
-#### 2. Прежде чем приступать к новому заданию, обновите `master`
-
-Обновить свой репозиторий из репозитория Академии можно так:
-
-```
-# В вашей локальной копии переключитесь в ветку master
-git checkout master
-
-# Заберите изменения из репозитория Академии¹
-git pull academy master
-
-# Отправьте изменения в ваш форк на Гитхабе
-git push
-```
-
-¹ В `academy` должна быть ссылка на репозиторий Академии. Если его там нет, добавьте:
-
-```
-git remote add academy git@github.com:htmlacademy-ecmascript/200558-big-trip-simple-19.git
-```
-
-Когда вы обновили `master`, создайте ветку для нового задания:
-
-```
-git checkout -b module2-task1
-```
-
-`module2-task1` — это название ветки. Под описанием каждого задания в интерфейсе курса для вас будет указано правильное название ветки.
-
---
-
-#### Есть вопрос?
-
-Посмотрите [коллекцию часто задаваемых вопросов по Git](http://firstaidgit.ru).
