+++ markup/img/icons/restaurant.png
