+++ markup/img/icons/ship.png
