+++ public/index.html
@@ -39,10 +39,11 @@
             <h2 class="visually-hidden">Trip events</h2>
 
             <!-- Сортировка -->
-
+                   
             <!-- Контент -->
           </section>
       </div>
     </main>
+    <script src="../bundle.js"></script>
   </body>
 </html>
