+++ requirements.md
@@ -1,6 +1,6 @@
 # Список внешних зависимостей
 
-Ниже представлен полный список пакетов, которые используются для разработки проекта «Большое путешествие». Обратите внимание, все пакеты должны быть установлены определённой версии. Версия пакета вводится после символа `@`. Например:
+Ниже представлен полный список пакетов, которые используются для разработки проекта «Менеджер задач». Обратите внимание, все пакеты должны быть установлены определённой версии. Версия пакета вводится после символа `@`. Например:
 
 ```bash
 npm i <package>@<version> -E
@@ -27,3 +27,4 @@ npm i eslint@8.28.0 -DE
 - babel-loader@9.1.0
 - css-loader@6.7.2
 - style-loader@3.3.1
+
