+++ src/render.js
@@ -5,15 +5,14 @@ const RenderPosition = {
   AFTEREND: 'afterend',
 };
 
-function createElement(template) {
+export function createElement(template) {
   const newElement = document.createElement('div');
   newElement.innerHTML = template;
 
   return newElement.firstElementChild;
 }
 
-function render(component, container, place = RenderPosition.BEFOREEND) {
+export function render(component, container, place = RenderPosition.BEFOREEND) {
   container.insertAdjacentElement(place, component.getElement());
 }
 
-export {RenderPosition, createElement, render};
