+++ src/view/create-waypoint-form-template.js
@@ -1,6 +1,7 @@
-import render from "../render.js";
-function creatingWaypointTemplate()  {
-  return `<form class="event event--edit" action="#" method="post">
+import {createElement} from '../render.js';
+function creatingWaypointTemplate() {
+  return `<li class="trip-events__item">
+              <form class="event event--edit" action="#" method="post">
                 <header class="event__header">
                   <div class="event__type-wrapper">
                     <label class="event__type  event__type-btn" for="event-type-toggle-1">
@@ -39,7 +40,7 @@ function creatingWaypointTemplate()  {
                         </div>
 
                         <div class="event__type-item">
-                          <input id="event-type-flight-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="flight" checked>
+                          <input id="event-type-flight-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="flight" checked="">
                           <label class="event__type-label  event__type-label--flight" for="event-type-flight-1">Flight</label>
                         </div>
 
@@ -76,7 +77,7 @@ function creatingWaypointTemplate()  {
                   <div class="event__field-group  event__field-group--time">
                     <label class="visually-hidden" for="event-start-time-1">From</label>
                     <input class="event__input  event__input--time" id="event-start-time-1" type="text" name="event-start-time" value="19/03/19 00:00">
-                    &mdash;
+                    —
                     <label class="visually-hidden" for="event-end-time-1">To</label>
                     <input class="event__input  event__input--time" id="event-end-time-1" type="text" name="event-end-time" value="19/03/19 00:00">
                   </div>
@@ -84,7 +85,7 @@ function creatingWaypointTemplate()  {
                   <div class="event__field-group  event__field-group--price">
                     <label class="event__label" for="event-price-1">
                       <span class="visually-hidden">Price</span>
-                      &euro;
+                      €
                     </label>
                     <input class="event__input  event__input--price" id="event-price-1" type="text" name="event-price" value="">
                   </div>
@@ -98,19 +99,19 @@ function creatingWaypointTemplate()  {
 
                     <div class="event__available-offers">
                       <div class="event__offer-selector">
-                        <input class="event__offer-checkbox  visually-hidden" id="event-offer-luggage-1" type="checkbox" name="event-offer-luggage" checked>
+                        <input class="event__offer-checkbox  visually-hidden" id="event-offer-luggage-1" type="checkbox" name="event-offer-luggage" checked="">
                         <label class="event__offer-label" for="event-offer-luggage-1">
                           <span class="event__offer-title">Add luggage</span>
-                          &plus;&euro;&nbsp;
+                          +€&nbsp;
                           <span class="event__offer-price">30</span>
                         </label>
                       </div>
 
                       <div class="event__offer-selector">
-                        <input class="event__offer-checkbox  visually-hidden" id="event-offer-comfort-1" type="checkbox" name="event-offer-comfort" checked>
+                        <input class="event__offer-checkbox  visually-hidden" id="event-offer-comfort-1" type="checkbox" name="event-offer-comfort" checked="">
                         <label class="event__offer-label" for="event-offer-comfort-1">
                           <span class="event__offer-title">Switch to comfort class</span>
-                          &plus;&euro;&nbsp;
+                          +€&nbsp;
                           <span class="event__offer-price">100</span>
                         </label>
                       </div>
@@ -119,7 +120,7 @@ function creatingWaypointTemplate()  {
                         <input class="event__offer-checkbox  visually-hidden" id="event-offer-meal-1" type="checkbox" name="event-offer-meal">
                         <label class="event__offer-label" for="event-offer-meal-1">
                           <span class="event__offer-title">Add meal</span>
-                          &plus;&euro;&nbsp;
+                          +€&nbsp;
                           <span class="event__offer-price">15</span>
                         </label>
                       </div>
@@ -128,7 +129,7 @@ function creatingWaypointTemplate()  {
                         <input class="event__offer-checkbox  visually-hidden" id="event-offer-seats-1" type="checkbox" name="event-offer-seats">
                         <label class="event__offer-label" for="event-offer-seats-1">
                           <span class="event__offer-title">Choose seats</span>
-                          &plus;&euro;&nbsp;
+                          +€&nbsp;
                           <span class="event__offer-price">5</span>
                         </label>
                       </div>
@@ -137,7 +138,7 @@ function creatingWaypointTemplate()  {
                         <input class="event__offer-checkbox  visually-hidden" id="event-offer-train-1" type="checkbox" name="event-offer-train">
                         <label class="event__offer-label" for="event-offer-train-1">
                           <span class="event__offer-title">Travel by train</span>
-                          &plus;&euro;&nbsp;
+                          +€&nbsp;
                           <span class="event__offer-price">40</span>
                         </label>
                       </div>
@@ -159,9 +160,10 @@ function creatingWaypointTemplate()  {
                     </div>
                   </section>
                 </section>
-              </form>`;
+              </form>
+            </li>`;
 }
-class creatingWaypoint {
+class CreatingWaypoint {
   getTemplate() {
     return creatingWaypointTemplate();
   }
@@ -178,3 +180,4 @@ class creatingWaypoint {
     this.element = null;
   }
 }
+export default CreatingWaypoint;
