+++ src/view/filters-template.js
@@ -1,4 +1,4 @@
-import {createElement} from "../render.js";
+import {createElement} from '../render.js';
 function filtersTemplate() {
   return `<div class="trip-filters__filter">
                   <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything" checked="">
@@ -6,7 +6,7 @@ function filtersTemplate() {
           </div>`;
 }
 class Filters {
-    getTemplate() {
+  getTemplate() {
     return filtersTemplate();
   }
